@@ -39,14 +39,14 @@
     "api": "node ./swagger/codegen.v3.js"
   },
   "dependencies": {
-    "@angular/animations": "~8.2.11",
-    "@angular/common": "~8.2.11",
-    "@angular/compiler": "~8.2.11",
-    "@angular/core": "~8.2.11",
-    "@angular/forms": "~8.2.11",
-    "@angular/platform-browser": "~8.2.11",
-    "@angular/platform-browser-dynamic": "~8.2.11",
-    "@angular/router": "~8.2.11",
+    "@angular/animations": "~9.0.7",
+    "@angular/common": "~9.0.7",
+    "@angular/compiler": "~9.0.7",
+    "@angular/core": "~9.0.7",
+    "@angular/forms": "~9.0.7",
+    "@angular/platform-browser": "~9.0.7",
+    "@angular/platform-browser-dynamic": "~9.0.7",
+    "@angular/router": "~9.0.7",
     "@antv/data-set": "^0.10.2",
     "@antv/g2": "^3.5.11",
     "@antv/g2-plugin-slider": "^2.1.1",
@@ -71,16 +71,16 @@
     "ngx-tinymce": "^7.0.0",
     "oidc-client": "^1.10.1",
     "qrious": "^4.0.2",
-    "rxjs": "~6.4.0",
+    "rxjs": "~6.5.4",
     "screenfull": "^5.0.0",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.14",
-    "@angular/cli": "~8.3.14",
-    "@angular/compiler-cli": "~8.2.11",
-    "@angular/language-service": "~8.2.11",
+    "@angular-devkit/build-angular": "~0.900.7",
+    "@angular/cli": "~9.0.7",
+    "@angular/compiler-cli": "~9.0.7",
+    "@angular/language-service": "~9.0.7",
     "@angularclass/hmr": "^2.1.3",
     "@datorama/akita-ngdevtools": "^4.0.0",
     "@delon/testing": "^8.8.0",
@@ -88,10 +88,10 @@
     "@types/jasminewd2": "~2.0.3",
     "@types/jszip": "^3.1.6",
     "@types/mockjs": "^1.0.2",
-    "@types/node": "~8.9.4",
+    "@types/node": "^12.11.1",
     "antd-theme-generator": "^1.1.7",
     "codecov": "^3.6.1",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^5.1.2",
     "gh-pages": "^2.1.1",
     "husky": "^3.0.9",
     "jasmine-core": "~3.4.0",
@@ -119,7 +119,7 @@
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "~3.5.3",
+    "typescript": "~3.7.5",
     "webpack-bundle-analyzer": "^3.6.0",
     "xlsx": "^0.15.1"
   },
