@@ -19,7 +19,7 @@ import { AuthService } from 'src/store/auth.service';
 })
 
 export class CoreModule {
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<CoreModule> {
     return {
       ngModule: CoreModule,
       providers: [
