@@ -66,7 +66,7 @@ export class DelonModule {
     throwIfAlreadyLoaded(parentModule, 'DelonModule');
   }
 
-  static forRoot(): ModuleWithProviders {
+  static forRoot(): ModuleWithProviders<DelonModule> {
     return {
       ngModule: DelonModule,
       providers: [...REUSETAB_PROVIDES, ...GLOBAL_CONFIG_PROVIDES],
